Default logic is to escalate to challenge if authentication failed. OnChallenge handler will then try to modify response object, throwing exception. Instead, set exception to known exception type if known authentication failure (used Maskinporten token on Altinn token endpoint) and use it to short-circuit onchallenge handler to set the response there instead.
