+++ src/Altinn.Broker.API/Models/Maskinporten/MaskinportenSecurityTokenException.cs
@@ -1,0 +1,5 @@
+﻿namespace Altinn.Broker.API.Models.Maskinporten;
+
+public sealed class MaskinportenSecurityTokenException : Exception
+{
+}
